Common data model
A model is something shared. It describes a common interface or protocol all
participants have to adhere to.

This commit therefore consolidates code hosters' repo related data type
declarations in a common module to ensure all drivers use compatible interfaces
and produce same output for consumption by crawler.
